+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Enabled pagination of search results.
+- Added `/api/v1/mutes` Mastodon API endpoint.
 
 ### Changed